fix int8 type issue

+++ aten/src/ATen/native/mkldnn/MKLDNNCommon.cpp
@@ -114,7 +114,7 @@ ideep::tensor itensor_from_tensor(const at::Tensor& tensor) {
   }
 }
 
-ideep::scale_t ConvertScales(const std::vector<double> &scales_z) {
+ideep::scale_t ConvertScales(const std::vector<float> &scales_z) {
   ideep::scale_t scales(scales_z.size());
 #ifdef _OPENMP
   #pragma omp parallel for schedule(static)
